Refactor test helper as detached functions
