+++ aquadoggo/src/db/stores/test_utils/mod.rs
@@ -5,7 +5,8 @@ mod runner;
 mod store;
 
 pub use helpers::{
-    doggo_test_fields, encode_entry_and_operation, insert_entry_operation_and_view, test_key_pairs,
+    add_document, add_schema, doggo_test_fields, encode_entry_and_operation,
+    insert_entry_operation_and_view, test_key_pairs,
 };
 pub use runner::{
     test_db, test_db_config, with_db_manager_teardown, TestDatabaseManager, TestDatabaseRunner,
