+++ aquadoggo/src/graphql/client/dynamic_query.rs
@@ -371,7 +371,9 @@ mod test {
     use rstest::rstest;
     use serde_json::json;
 
-    use crate::db::stores::test_utils::{test_db, TestDatabase, TestDatabaseRunner};
+    use crate::db::stores::test_utils::{
+        add_document, add_schema, test_db, TestDatabase, TestDatabaseRunner,
+    };
     use crate::test_helpers::graphql_test_client;
 
     #[rstest]
@@ -382,18 +384,22 @@ mod test {
             let key_pair = random_key_pair();
 
             // Add schema to node.
-            let schema = db
-                .add_schema("schema_name", vec![("bool", FieldType::Bool)], &key_pair)
-                .await;
+            let schema = add_schema(
+                &mut db,
+                "schema_name",
+                vec![("bool", FieldType::Bool)],
+                &key_pair,
+            )
+            .await;
 
             // Publish document on node.
-            let view_id = db
-                .add_document(
-                    schema.id(),
-                    vec![("bool", true.into())].try_into().unwrap(),
-                    &key_pair,
-                )
-                .await;
+            let view_id = add_document(
+                &mut db,
+                schema.id(),
+                vec![("bool", true.into())].try_into().unwrap(),
+                &key_pair,
+            )
+            .await;
             let document_id =
                 DocumentId::from(view_id.graph_tips().first().unwrap().as_hash().to_owned());
 
@@ -409,7 +415,7 @@ mod test {
                 }}
             }}"#,
                 type_name = schema.id().to_string(),
-                view_id = view_id.to_string(),
+                view_id = view_id,
                 document_id = document_id.as_str()
             );
 
@@ -512,13 +518,18 @@ mod test {
             let key_pair = random_key_pair();
 
             // Add schema to node.
-            let schema = db
-                .add_schema("schema_name", vec![("bool", FieldType::Bool)], &key_pair)
-                .await;
+            let schema = add_schema(
+                &mut db,
+                "schema_name",
+                vec![("bool", FieldType::Bool)],
+                &key_pair,
+            )
+            .await;
 
             // Publish document on node.
-            db.add_document(
-                &schema.id(),
+            add_document(
+                &mut db,
+                schema.id(),
                 vec![("bool", true.into())].try_into().unwrap(),
                 &key_pair,
             )
